requirements.txt
numpy==1.24.3
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
cnocr==2.2.2.3
torch==2.0.1
Pillow==9.5.0
pdfplumber==0.9.0
PyMuPDF==1.22.5
